replacing end chests with regular chests at death
working on replacing contents, still very buggy

+++ AlmostHardcoreMinecraft/src/org/c0z3n/AlmostHardcore.java
@@ -14,6 +14,7 @@ import org.bukkit.Material;
 import org.bukkit.World;
 import org.bukkit.block.Biome;
 import org.bukkit.block.Block;
+import org.bukkit.block.Chest;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
@@ -25,8 +26,11 @@ import org.bukkit.event.entity.PlayerDeathEvent;
 import org.bukkit.event.player.PlayerBedEnterEvent;
 import org.bukkit.event.player.PlayerJoinEvent;
 import org.bukkit.event.player.PlayerRespawnEvent;
+import org.bukkit.inventory.Inventory;
+import org.bukkit.inventory.ItemStack;
 import org.bukkit.plugin.java.JavaPlugin;
 import org.bukkit.scheduler.BukkitScheduler;
+
 
 public class AlmostHardcore extends JavaPlugin{	
 	private com.avaje.ebean.EbeanServer db;
@@ -54,7 +58,34 @@ public class AlmostHardcore extends JavaPlugin{
 				hardcorePlayer hcp = db.find(hardcorePlayer.class).where().eq("id", player.getUniqueId()).findUnique();
 				hcp.addDeath();
 				hcp.setNightsAlive(0);
-				updateGlobalSpawnLocation(player);
+//				updateGlobalSpawnLocation(player);
+				ItemStack[] endChestInv = player.getEnderChest().getContents();
+				int numEnderChests = db.find(hardcoreEnderChest.class).where().eq("owner", player.getUniqueId()).findRowCount();
+				List<hardcoreEnderChest> enderChestList = db.find(hardcoreEnderChest.class).where().eq("owner", player.getUniqueId()).findList();
+				ItemStack[][] chunkedEndChestInv = splitInventory(endChestInv, numEnderChests);
+				int idx = 0;
+                for (hardcoreEnderChest endChest : enderChestList){
+                	World world = getServer().getWorlds().get(0);
+                    int bX = (int)endChest.getX();
+                	int bY = (int)endChest.getY();
+                	int bZ = (int)endChest.getZ();
+                	Block chestBlock = world.getBlockAt(bX,bY,bZ);
+                	Material convertToMaterial = Material.TRAPPED_CHEST;
+                	if(world.getBlockAt(bX + 1, bY, bZ).getType() != Material.CHEST && world.getBlockAt(bX - 1, bY, bZ).getType() != Material.CHEST && world.getBlockAt(bX, bY, bZ + 1).getType() != Material.CHEST && world.getBlockAt(bX, bY, bZ - 1).getType() != Material.CHEST){
+                		convertToMaterial = Material.CHEST;
+                	}
+                	chestBlock.setType(convertToMaterial);
+                	
+                	Block newChestBlock = world.getBlockAt(bX,bY,bZ);
+                    if(newChestBlock.getState() instanceof Chest){
+                        Chest chest = (Chest) newChestBlock.getState();
+                        Inventory chestInv = chest.getInventory();
+//                        chestInv.setContents(chunkedEndChestInv[idx]);
+                        chestInv.setContents(endChestInv);
+                    }
+                    db.delete(endChest);
+                	idx = idx + 1;
+                }
 				player.getEnderChest().clear();
 				db.save(hcp);
 			}
@@ -101,7 +132,7 @@ public class AlmostHardcore extends JavaPlugin{
 			
 			@EventHandler
 			public void onBlockBreak(BlockBreakEvent e) {
-				if ( e.getBlock().getType() == Material.ENDER_CHEST){
+				if ( e.getBlock().getType() == Material.ENDER_CHEST && db.find(hardcoreEnderChest.class).where().eq("x", e.getBlock().getX()).eq("y", e.getBlock().getY()).eq("z", e.getBlock().getZ()).findRowCount() > 0){
 					hardcoreEnderChest endChest = db.find(hardcoreEnderChest.class).where().eq("x", e.getBlock().getX()).eq("y", e.getBlock().getY()).eq("z", e.getBlock().getZ()).findUnique();
 					db.delete(endChest);
 				}
@@ -204,6 +235,23 @@ public class AlmostHardcore extends JavaPlugin{
         list.add(hardcoreSpawn.class);
         list.add(hardcoreEnderChest.class);
         return list;
+    }
+    
+    private ItemStack[][] splitInventory(ItemStack[] inventory, int numChunks) {
+    	
+        int chunkSize = (int)Math.ceil((double)inventory.length / numChunks);
+        ItemStack[][] output = new ItemStack[numChunks][];
+
+        for(int i = 0; i < numChunks; ++i) {
+            int start = i * chunkSize;
+            int length = Math.min(inventory.length - start, chunkSize);
+
+            ItemStack[] temp = new ItemStack[length];
+            System.arraycopy(inventory, start, temp, 0, length);
+            output[i] = temp;
+        }
+
+        return output;
     }
 	
 	private int randCoord() {
